Add support for Azure AD with User Password
I struggled with this for the better part of a half of a day, hopefully this can be pulled in to save someone the trouble I went through.

Issue is that when using a service principal azure AD operations will work as the auth flow is configured for that eventuality. When using username password, it is NOT. This fixes that issue. 

I would guess that a test should be written for this case so that it is dealt with in the future, but I am probably not the person to add a test to your setup.