Ansible.builtin.changed (#2003)
* Update assert tests to use is changed/failed module

Ansible 2.19 is stricter about evaluating booleans and since there is a
changed or failed module to assert against it makes sense to move to
that and be consistent.

* Update tests to be consistent in changed and failed

Ansible 2.19 is more strict about checks.  Let's standardize on
the ansible.builtin.changed and ansible.builtin.failed modules for
checking each condition.

https://docs.ansible.com/ansible/latest/collections/ansible/builtin/changed_test.html
https://docs.ansible.com/ansible/latest/collections/ansible/builtin/failed_test.html