Fix exception handling in azure_rm_arcssh
When setting up the ARC proxy on Azure we need to check if the relay
endpoint exists.  If it doesn't exist we rely on ResourceNotFoundError
exception to know that we need to create one.

This adds an error_map where you can optionally raise specific exceptions
based on http status codes while still supporting the exisiting code.